// ARSIP/src/packages/1527_lab_order_details/models/labOrderDetailResponse.ts
export interface LabOrderDetailResponse {
  status: number;
  message: string;
  data: LabOrderDetailItem[];
}

export interface LabOrderDetailItem {
  id: string;
  id_ref_lab_order: string;
  id_ref_order_jenis_pasien: string;
  id_episode: string;
  id_kel: string;
  id_reg: string;
  id_order_poli_details: string;
  id_order: string;
  id_visit: string;
  id_pasien_episode: string;
  id_dept_ori: string;
  id_dept_dest: string;
  id_rs_antrean_cs: string;
  id_order_poli: string;
  id_ref_tarif_group: string;
  id_ref_kegiatan: string;
  id_group_ref_kegiatan: string;
  id_parent_ref_kegiatan: string;
  created_date: string;
  input_date: string;
  preapproved_date: string;
  approved_date: string;
  deptori_nama_dept_ori: string;
  nama_pasien: string;
  alamat_pasien: string;
  status_pasien: string;
  deptori_nama_dept_dest: string;
  hasillab_klinis_pasien: string;
  created_by: string;
  orderp_no_order: string;
  cito: string;
  no_rm: string;
  no_reg: string;
  catatan: string;
  kesan: string;
  pesan: string;
  selesai_pemeriksaan: string;
  sample_diterima: string;
  analis_nama: string;
  analis_no_rm: string;
  dpjp_lab_nama: string;
  dpjp_lab_no_rm: string;
  pj_lab_nama: string;
  pj_lab_no_rm: string;
  dpjp_nip: string;
  xstatus: string;
  preapproved_by: string;
  approved_by: string;
  tarif_penjamin: string;
  tarif_kelas: string;
  orderpd_nama_tarif: string;
  nama_kegiatan: string;
  harga_normal: string;
  harga_tarif: string;
  harga_satuan: string;
  harga_up: string;
  harga_diskon: string;
  tarif_group: string;
  tarif_jenis_rawat: string;
  resume_billing_group: string;
  kode_transaksi_pembayaran: string;
  cara_pembayaran: string;
  tipe_lab_order: string;
  sumber_test_covid: string;
  tipe_test_covid: string;
  tarif_excess: string;
  tanggal_data_order_terkirim_ke_lis: string;
}